Rename pages/demo to pages/components; DemoIndex -> ComponentsIndex
