+++ packages/playground/src/pages/components/AvatarDemo.tsx
