+++ packages/playground/src/pages/components/Example.module.scss
