+++ packages/playground/src/pages/components/prismGlobal.ts
